Create sample_queries.txt (#5)
The goal of this file is to create a set of theories we can reflect back
on when making tweaks and judge behavior. Its worth noting that almost
all of these have a bias coming from the user and the goal is the AI
doesn't give in to it. Some of these are just pure nonsense.

This is probably too many to worry about in the hackathon, but I'm just
jotting down ideas. If we want to reduce this down to much fewer
theories, then I think we just want to consider:
- A nonsense query
- A perfectly reasonable question that the AI could probably answer
without research
- A reasonable question the AI could research
- A question thats very involved and requires a lot of research
- A biased query
- A query that comes in with a lot of information
- A query that comes in with a lot of misinformation
- A loaded question