Super agi toolkit and codespaces set-up (#10)
This is the pre-work that investigated running with SuperAGI. It makes
sense to build off of this. It already has run profiles so its just a
matter of running the "Run backend and celery" compound and also running
the "Start Super AGI docker containers" tasks.

The toolkit is a placeholder and has no value other than showing how we
would do that.