Deep research tool (#12)
This is a basic foundation. The AI will recognize it needs to do
research and will use the tool. The tool extracts up to 5 topics (so far
its always been 5), and then is instructed to go use the Deep Research
Summary Tool. The AI agent will use this tool, which returns markdown.
If using a goal orientated agent, I noticed it fails to recognize that
its done, so I added something about it using the finish tool.


![work-so-far-screen](https://github.com/JamesHutchison/research-agi-agent/assets/122519877/4c539476-7888-46c3-a3bb-ef6ea948d54e)