Deep research topic investigator (#13)
This adds an non-AGI subagent to do topic research. The logic is
currently simple, it does a google search, grabs the first 3k characters
for the top three results. It writes this to a file for later recall.
Then it succesfully determines its done and writes the output.

I highly suggest starting from the output sample when developing the
summary writer to save yourself time and money